refactor : PathService에 있는 비즈니스 로직 코드 Path 도메인으로 이동
